finished Results page/ added back to top button/ created backToTopIcon comp/ general adjustments
finished results page and added back to top component button and also provided a fix for the search bar search icon, will most likely further correct it more. Created icon for button to go back to top of results page so now the page is ready for infinite scrolling with some mild caveats.